Священные слова "Москва за нами!", Мы помним со времён Бородина.
