+++ src/func.rs
@@ -1,39 +1,23 @@
-#![allow(unused_imports)]
-#![allow(dead_code)]
-#![allow(unused_variables)]
-
-use std::any::TypeId;
 use std::marker::PhantomData;
 use std::sync::atomic::Ordering::SeqCst;
 
-use crate::pylon::{DynBase, GcInfo, lifetime_check, Ptr, VMPtrToRust};
+use crate::data::{DynBase, Ptr, GcInfo};
 use crate::tyck::{StaticBase, TypeCheckInfo};
-
-pub enum Storage {
-    VMOwned,
-    SharedWithHost,
-    MutSharedWithHost,
-}
-
-pub enum RustArgLifetime {
-    Move,
-    Copy,
-    Share,
-    MutShare,
-}
+use crate::cast::{RustLifetime, lifetime_check};
+use crate::cast::into::VMPtrToRust;
 
 pub trait RustCallable<'a> {
     fn is_unsafe(&self) -> bool;
 
-    fn param_specs(&self) -> Vec<(TypeCheckInfo, RustArgLifetime)>;
+    fn param_specs(&self) -> Vec<(TypeCheckInfo, RustLifetime)>;
 
-    fn return_value_spec(&self) -> (TypeCheckInfo, RustArgLifetime);
+    fn return_value_spec(&self) -> (TypeCheckInfo, RustLifetime);
 
     unsafe fn call_prechecked(&self, args: &'a [Ptr<'a>]) -> Result<Ptr<'a>, String>;
 
     fn call(&self,
             args: &'a [Ptr<'a>],
-            ret_tyck_info: Option<TypeCheckInfo>)
+            _ret_tyck_info: Option<TypeCheckInfo>)
         -> Result<Ptr<'a>, String>
     {
         let param_specs = self.param_specs();
@@ -75,19 +59,19 @@ impl<'a, A, B, RET, FN> RustCallable<'a> for RustCallBind2<A, B, RET, FN>
         false
     }
 
-    fn param_specs(&self) -> Vec<(TypeCheckInfo, RustArgLifetime)> {
+    fn param_specs(&self) -> Vec<(TypeCheckInfo, RustLifetime)> {
         vec![
             (A::tyck_info(), A::lifetime_info()),
             (B::tyck_info(), B::lifetime_info())
         ]
     }
 
-    fn return_value_spec(&self) -> (TypeCheckInfo, RustArgLifetime) {
+    fn return_value_spec(&self) -> (TypeCheckInfo, RustLifetime) {
         (RET::tyck_info(), RET::lifetime_info())
     }
 
     unsafe fn call_prechecked(&self, args: &'a [Ptr<'a>]) -> Result<Ptr<'a>, String> {
-        let ret: RET = self.inner.call((
+        let _ret: RET = self.inner.call((
             <() as VMPtrToRust<'a, A>>::any_cast(args[0].clone())?
             , <() as VMPtrToRust<'a, B>>::any_cast(args[1].clone())?
         ));
